Add files via upload
Street Surfer Cat is an endless runner game inspired by the popular Subway Surfers. In this game, players control a sprightly cat character as it dashes through bustling streets, dodging obstacles, and collecting various power-ups. The objective is to achieve the highest score by running as far as possible while avoiding collisions with obstacles and enemies.